+++ scripts/sql/tables/passport.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `passport` (
-                            `id` INT(11) NOT NULL,
+                            `id` BIGINT NOT NULL,
                             `name` VARCHAR(255) NOT NULL,
                             `password` VARCHAR(255) NOT NULL,
                             `created_at` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
